__CRAZY=Yes cleanup… can't they even use compiler warnings… cruel netbsd code *sigh*

+++ src/usr.sbin/makefs/nbsrc/sys/ufs/ffs/fs.h
@@ -679,9 +679,9 @@ struct ocg {
  * Determining the size of a file block in the file system.
  */
 #define	blksize(fs, ip, lbn) \
-	(((lbn) >= NDADDR || (ip)->i_size >= lblktosize(fs, (lbn) + 1)) \
-	    ? (fs)->fs_bsize \
-	    : (fragroundup(fs, blkoff(fs, (ip)->i_size))))
+	(((lbn) >= NDADDR || (off_t)(ip)->i_size >= lblktosize(fs, (lbn) + 1)) \
+	    ? (int32_t)(fs)->fs_bsize \
+	    : (int32_t)(fragroundup(fs, blkoff(fs, (ip)->i_size))))
 
 #define sblksize(fs, size, lbn) \
 	(((lbn) >= NDADDR || (size) >= ((lbn) + 1) << (fs)->fs_bshift) \
